Add fallback model support and warning mechanism for unsupported Bedrock models (#1224)
<!--
Thank you for opening a pull request!

Please add a brief description of the proposed change here.
Also, please tick the appropriate points in the checklist below.
-->

## Motivation and Context
<!-- Why is this change needed? What problem does it solve? -->
[KG-595](https://youtrack.jetbrains.com/issue/KG-595) Relax
getBedrockModelFamily error to a warning when using custom models

When the user adds a Bedrock model from a family that's not added to the
list we have, it's good to have a fallback mechanism.
E.g. the user knows that the LLM API of a new model works exactly like
the Anthropic one. In this case, they'll get a warning and can continue
using the fallback Anthropic API until their own serializer is not
added.

Source:
https://github.com/JetBrains/koog/issues/1210#issuecomment-3601786108

## Breaking Changes
<!-- Will users need to update their code or configurations? -->

---

#### Type of the changes
- [x] New feature (non-breaking change which adds functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Tests improvement
- [x] Refactoring

#### Checklist
- [x] The pull request has a description of the proposed change
- [x] I read the [Contributing
Guidelines](https://github.com/JetBrains/koog/blob/main/CONTRIBUTING.md)
before opening the pull request
- [x] The pull request uses **`develop`** as the base branch
- [x] Tests for the changes have been added
- [x] All new and existing tests passed

##### Additional steps for pull requests adding a new feature
- [x] An issue describing the proposed change exists
- [x] The pull request includes a link to the issue
- [ ] The change was discussed and approved in the issue
- [x] Docs have been added / updated